Issue were ph would be saved as the rounded whole number has been solved

+++ blueflamingoapi/models/ph.py
@@ -3,5 +3,5 @@ from django.db import models
 
 class Ph(models.Model):
     """Tag Model"""
-    ph = models.IntegerField()
+    ph = models.FloatField()
     message = models.CharField(max_length=175)
